Added code that checks if list of boolean has all the same values, and if yes, then the list is reduce to one element with this value.  If the list of values has both True and False values, we do not modify it.
This change affects the 'Has Relationships' property.

If the list of booleans for this property has all values that are equal, for example,
[True, True, True, True], then we replace it with [True] and show to the user only one true
instead of four.

On the other hand, if the list of booleans for this property has both True and False values,
say, [True, False, True, True], then this list will be added to CX model as is and shown to user like

Has Relationships:
true
false
true
true